Tests/LibWeb/Text/expected/HTML/parser-script-source-lines.txt

HTML split start tag line: 17
SVG split start tag line: 19
